requirements.txt
customtkinter==5.2.2
pillow==10.2.0
openpyxl==3.1.2
pyinstaller==6.4.0
pandas==2.2.1
numpy==1.26.4
python-dateutil==2.9.0.post0 